Separate VEP code. Add tests.
