+++ src/genophenocorr/preprocessing/test_data/vep_response/README.md
@@ -4,7 +4,7 @@ The folder contains JSON files with example VEP responses.
 
 ## `missense.json`
 
-QUERY:
+Corresponds to variant in `../misssense_test.json`.
 ```shell
 curl 'https://rest.ensembl.org/vep/human/region/16:89279135-89279135/C?LoF=1&canonical=1&domains=1&hgvs=1&mutfunc=1&numbers=1&protein=1&refseq=1&mane=1&transcript_version=1&variant_class=1' \
 -H 'Content-type:application/json' | python3 -m json.tool > missense.json
@@ -12,7 +12,9 @@ curl 'https://rest.ensembl.org/vep/human/region/16:89279135-89279135/C?LoF=1&can
 
 ## `deletion.json`
 
+Corresponds to variant in `../deletion_test.json`.
+
 ```shell
-curl 'https://rest.ensembl.org/vep/human/region/16:89279135-89279135/C?LoF=1&canonical=1&domains=1&hgvs=1&mutfunc=1&numbers=1&protein=1&refseq=1&mane=1&transcript_version=1&variant_class=1' \
+curl 'https://rest.ensembl.org/vep/human/region/16:89284129-89284134/C?LoF=1&canonical=1&domains=1&hgvs=1&mutfunc=1&numbers=1&protein=1&refseq=1&mane=1&transcript_version=1&variant_class=1' \
 -H 'Content-type:application/json' | python3 -m json.tool > deletion.json
 ```
